feat(garmin): fix activity serialization and improve logging
Change duration field from int to float in GarminActivity model to support
fractional seconds. Add proper datetime serialization to ISO format for
JSON compatibility when sending activities to backend. Enhance logging with
detailed success/error information for activity sync operations.

+++ app/models/garmin_models.py
@@ -56,7 +56,7 @@ class GarminActivity(BaseModel):
     activity_type: str
     start_time: datetime
     end_time: Optional[datetime] = None
-    duration: Optional[int] = None  # в секундах
+    duration: Optional[float] = None  # в секундах
     distance: Optional[float] = None  # в метрах
     average_hr: Optional[int] = None
     max_hr: Optional[int] = None
